Reorganized actions and queries.
