+++ src/app/_components/create-transaction-dialog-wrapper.tsx
@@ -1,7 +1,7 @@
 'use server';
 
 import { CreateTransactionDialog } from '@/app/_components/create-transaction-dialog';
-import { getSimpleAccounts } from '@/lib/db/accounts';
+import { getSimpleAccounts } from '@/lib/db/accounts.queries';
 
 export async function CreateTransactionDialogWrapper() {
   const accounts = await getSimpleAccounts();
