+++ src/lib/db/currencies.queries.ts
@@ -1,10 +1,16 @@
+'use server';
+
 import { Currency, CurrencyMapper } from '@/lib/types/currencies.types';
 import { createServerSupabaseClient } from '@/lib/utils/supabase/server';
 
 export async function getCurrencies(): Promise<Currency[]> {
   const supabase = createServerSupabaseClient({ next: { revalidate: 60, tags: ['currencies'] } });
 
-  const { data } = await supabase.from('currencies').select();
+  const { data, error } = await supabase.from('currencies').select();
+
+  if (error) {
+    throw error;
+  }
 
   return data ?? [];
 }
@@ -12,7 +18,7 @@ export async function getCurrencies(): Promise<Currency[]> {
 export async function getCurrencyMapper(toCurrency: string): Promise<CurrencyMapper> {
   const supabase = createServerSupabaseClient({ next: { revalidate: 60, tags: ['currencies'] } });
 
-  const { data } = await supabase
+  const { data, error } = await supabase
     .from('exchange_rates')
     .select('from, rate')
     .eq('to', toCurrency)
@@ -20,6 +26,10 @@ export async function getCurrencyMapper(toCurrency: string): Promise<CurrencyMap
       ...response,
       data: Object.fromEntries(response.data?.map((item) => [item.from, item.rate]) ?? []),
     }));
+
+  if (error) {
+    throw error;
+  }
 
   return data;
 }
