+++ src/lib/types/transactions.types.ts
@@ -1,0 +1,8 @@
+export interface Transaction {
+  id: string;
+  subaccountId: string;
+  type: string;
+  amount: number;
+  balance: number;
+  startedDate: string;
+}
